@@ -5,26 +5,42 @@
   "main": "gulpfile.js",
   "scripts": {
     "build": "gulp build",
-    "watch": "gulp watch"
+    "watch": "gulp watch",
+    "fix": "eslint --fix --ext .js,.ts app/",
+    "test": "echo \"Error: 测试报错\" && exit 1"
+  },
+  "lint-staged": {
+    "app/**/*.{js,ts}": [
+      "eslint --fix",
+      "git add"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged",
+      "post-commit": "echo \"提交完毕\""
+    }
   },
   "author": "caico",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/preset-env": "^7.4.5",
-    "babel-eslint": "^10.0.1",
-    "colors": "^1.3.3",
+    "@babel/core": "^7.7.2",
+    "@babel/preset-env": "^7.7.1",
+    "@typescript-eslint/parser": "^2.8.0",
+    "babel-eslint": "^10.0.3",
+    "colors": "^1.4.0",
     "cssnano": "^4.1.10",
     "del": "^4.1.1",
+    "eslint": "^6.6.0",
     "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^9.2.0",
+    "eslint-plugin-standard": "^4.0.1",
     "gulp": "^4.0.2",
     "gulp-autoprefixer": "^6.1.0",
     "gulp-babel": "^8.0.0",
     "gulp-concat": "^2.6.1",
-    "gulp-imagemin": "^6.0.0",
+    "gulp-imagemin": "^6.2.0",
     "gulp-jshint": "^2.1.0",
     "gulp-postcss": "^8.0.0",
     "gulp-rename": "^1.4.0",
@@ -32,26 +48,33 @@
     "gulp-sourcemaps": "^2.6.5",
     "gulp-tslint": "^8.1.4",
     "gulp-uglify": "^3.0.2",
-    "jshint": "^2.10.2",
-    "node-sass": "^4.12.0"
+    "husky": "^2.4.1",
+    "jshint": "^2.10.3",
+    "lint-staged": "^9.4.3",
+    "node-sass": "^4.13.0"
   },
   "dependencies": {
     "@types/gulp": "^4.0.6",
     "@types/gulp-autoprefixer": "0.0.31",
     "@types/gulp-concat": "0.0.32",
+    "@types/gulp-imagemin": "^4.1.1",
     "@types/gulp-rename": "0.0.33",
     "@types/gulp-sass": "^4.0.0",
     "@types/gulp-sourcemaps": "0.0.32",
     "@types/gulp-uglify": "^3.0.6",
+    "@types/jsonfile": "^5.0.0",
     "@types/node-sass": "^4.11.0",
-    "eslint-plugin-promise": "^4.1.1",
+    "@types/yargs": "^13.0.3",
+    "@typescript-eslint/eslint-plugin": "^2.8.0",
+    "eslint-plugin-promise": "^4.2.1",
     "gulp-eslint": "^5.0.0",
     "gulp-typescript": "^5.0.1",
     "gulpclass": "^0.2.0",
+    "jsonfile": "^5.0.0",
     "miniprogram-api-typings": "^2.9.2",
-    "ts-node": "^8.3.0",
-    "typescript": "^3.5.1",
-    "yargs": "^13.2.4"
+    "ts-node": "^8.5.2",
+    "typescript": "^3.7.2",
+    "yargs": "^13.3.0"
   },
   "eslintIgnore": [
     "dist/**/*.js"
