@@ -11,12 +11,12 @@ Component({
     }
   },
   data: {
-    auctionList:[
-      {username:'小东'},
-      {username:'小东'},
-      {username:'小东'},
-      {username:'小东'},
-      {username:'小东'},
+    auctionList: [
+      { username: '小东' },
+      { username: '小东' },
+      { username: '小东' },
+      { username: '小东' },
+      { username: '小东' }
     ]
   }
 })
