@@ -1,15 +1,9 @@
 Page({
   data: {
-    auctionLixt:[
-      {photos:'翡翠玉佩'},
-      {photos:'翡翠玉佩翡翠玉佩翡翠玉佩'},
-      {photos:'翡翠玉佩'},
+    auctionLixt: [
+      { photos: '翡翠玉佩' },
+      { photos: '翡翠玉佩翡翠玉佩翡翠玉佩' },
+      { photos: '翡翠玉佩' }
     ]
   }
-  onLoad() {
-
-  },
-  onUnload() {
-
-  }
-});
+})
