@@ -1,17 +1,19 @@
 Page({
   data: {
-    titleList:[
-      {txtname:'动态'},
-      {txtname:'玩物日志'},
+    titleList: [
+      { txtname: '动态' },
+      { txtname: '玩物日志' }
     ],
-    selsectIndex:0,
-  }
-  selectBtn({currentTarget: {dataset}}){
-    const { index } = dataset;
+    selsectIndex: 0
+  },
+  selectBtn ({ currentTarget: {
+    dataset = { index: 0 }
+  }}) {
+    const { index } = dataset
     this.setData({
-      selsectIndex:index
+      selsectIndex: index
     })
-  }
+  },
   onShow (): void {
     if (typeof this.getTabBar === 'function' && this.getTabBar()) {
       this.getTabBar().setData({
