@@ -1,18 +1,5 @@
 Page({
   data: {
-    goodsList:[
-      {username:'小唐马'},
-      {username:'小唐马'},
-      {username:'小唐马'},
-      {username:'小唐马'},
-      {username:'小唐马'},
-      {username:'小唐马'},
-    ]
-  }
-  onLoad() {
-
-  },
-  onUnload() {
 
   }
-});
+})
